refactor(App): extract AppView component
