+++ frontend/src/views/Register.vue
@@ -1,39 +1,44 @@
 <template>
-  <v-container>
-    <v-row>
-      <v-col>
-        <h2>Register</h2>
-      </v-col>
-    </v-row>
-    <v-row>
-      <v-col>
-        <v-form>
-          <v-text-field
-            v-for="(value, name) in form"
-            :id="`id_${name}`"
-            :key="name"
-            v-model="value.value"
-            :label="value.label"
-            :type="value.type"
-          />
-          <v-btn
-            id="submit_button"
-            type="submit"
-            @click="submit"
-          >
-            Register
-          </v-btn>
-        </v-form>
-      </v-col>
-    </v-row>
-  </v-container>
+  <app-view>
+    <v-container>
+      <v-row>
+        <v-col>
+          <h2>Register</h2>
+        </v-col>
+      </v-row>
+      <v-row>
+        <v-col>
+          <v-form>
+            <v-text-field
+              v-for="(value, name) in form"
+              :id="`id_${name}`"
+              :key="name"
+              v-model="value.value"
+              :label="value.label"
+              :type="value.type"
+            />
+            <v-btn
+              id="submit_button"
+              type="submit"
+              @click="submit"
+            >
+              Register
+            </v-btn>
+          </v-form>
+        </v-col>
+      </v-row>
+    </v-container>
+  </app-view>
 </template>
 
 <script>
 /** @file view for registering */
 
+import AppView from '../components/AppView.vue';
+
 export default {
   name: 'Register',
+  components: {AppView},
   data() {
     return {
       form: {
